% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_epigenome.R
\name{plot_epigenome}
\alias{plot_epigenome}
\title{A Function That Plots Roadmap Data}
\usage{
plot_epigenome(roadmap_matrices_metadata, matrices, marks, list_of_feat,
  feat_types, PLOT_TRSCR_TYPE = FALSE, PLOT_NORM_EXPR = FALSE,
  PLOT_SPERMATO_EXPR = FALSE, PLOT_CPG = FALSE, PLOT_METH = FALSE,
  cex = 2.5)
}
\arguments{
\item{roadmap_matrices_metadata}{metadata of roadmap matrices}

\item{matrices}{matrices to be plotted}

\item{marks}{marks of interest}

\item{list_of_feat}{list of index of gene (same index as roadmap matrices)}

\item{feat_types}{A indexed vector of types}

\item{PLOT_TRSCR_TYPE}{boolean specifyinf what to plot}

\item{PLOT_NORM_EXPR}{boolean specifyinf what to plot}

\item{PLOT_SPERMATO_EXPR}{boolean specifyinf what to plot}

\item{PLOT_CPG}{boolean specifyinf what to plot}

\item{PLOT_METH}{boolean specifyinf what to plot}

\item{cex}{grpahical parameter specifying size of characters}
}
\description{
This function plots roadmap data.
}

